Change Witch screen limit calculations
Collision between Witch and screen limits now uses collision rectangle.

Previously, this calculation only took into the account the size of the Witch texture, but it will now make use of the collision rectangle's dimensions instead.
It is also now possible to scale the size of the collision rectangle if desired.