Implement main Render Texture
The entire game is now drawn onto a render texture.

In order to make the movement as close to pixel perfect as possible, the game runs internally at a low resolution that fits the sprites. The problem is this causes the game window to be tiny and low res as well.
To solve this, we first render the game on a low resolution render texture, which is then drawn onto (and stretched to fit) a bigger window.