Implement GameState and PlayState
An abstract class called GameState has been created.

This class is used as a parent class for every class that defines a game state. It can be used to implement things like menus, pause screens, other levels, etc.
The current state of the game has now been implemented in a PlayState class that inherits from GameState.