Stabilize scroll animations and framerate
Target framerate and the speed of scrolling elements have been adjusted.

Some background layers had a visible judder when scrolling if the game ran at certain framerates. This happened because their speeds weren't multiples of the current framerate, which varies depending on monitor.
To solve this, speed of the fastest scrolling elements has been set to a multiple of 60.
The target framerate is now set to a multiple of 60 taking into account the monitor's refresh rate.
The game now runs with VSync on in Borderless Window mode.